Remove PayPalService and its state vars.
